lorin / sean: I added a bool function to the units to check if they have moved that turn already or not, it's called hasMoved() make sure you add in a segment to check that before a unit can be moved. Also, it's important to note that any unit that starts its turn in enemy territory can't move that turn, so set it's has moved to true with unit->setMoved(); (no parameter neccessary)
sean: 
I began to mod the army class in the loader and I think you can see what I'm going for. 
need seperate arrays for the different positions of the units. 
first is a 2d array to hold all units on the field. needs to be big enough to hold the max
	number of units the player could possibly be on the field (ie all units and reinforcements)
second is a 2d for all units that have been exited off the map, same size as above
third is a 2d for all units that have been killed, same size as above
fourth is complicated. its a 3d for reinforcments the idea is it's an array of arrays of pointers
	the first layer is which reinforcement it is, the second is an array of what is being reinforced
	so like if I get three sets of reinforcements, each of five units it'll need to be a unitClass[3][5]* right
	but it needs to be dynamic to what is being reinforced
so I could end up w/ unitClass[3]** with unitClass[0] = unitClass[5]* and unitClass[1] = unitClass[2]*
am I explaining this right? skype me if you got questions.
I also added a string to the units variables to store its name.

lorin: 
noticed a bug where if a unit is stepping off of a road onto a non road it treats that unit as 
if its movement cost is 1. so when I go from a tile with a road to a forest tile w/o road it only costs
one move point.

lorin I need a function written for the map like so bool openRoad(int player, int numStartNodes, int numExitNodes,
map_node ** startNodes, map_node ** exitNodes);
that searches for an open road from any of the start nodes to any of the end nodes. the player values
will be either 0 (clear of any unit) 1 (clear of union units) 2 (clear of confed units). I also need
to be able to pass it a node coordinates and check if there is a clear path to any exit node from there
doesn't have to be written now, just telling you what's up.

robert:
okay, I've split the game up into states with an in to track that state. if you need a new state for
something, just add it into the enum at the top of game data handler.h. inputhandler, draw, and update all
now have a switch(gameState) built into them, so just add it as a case.

everyone:
read the above one to note gameState.
need to come up with a way to better handle which "player" is the actual player. lorin, you and I 
need to sit down for like a day with this udp shit and restructure all this shit so that it will
be able to handle the networking. in it's current state it can't even come close.
I think it's best if robert and sean work on combat (in part so robert can get his hands into more
code) while lorin and I work together to hammer out the udp stuff.

lorin and I were talking and so far I think the best way to handle reinforcing of units is to have
a graphical representation of reinforcable units that can be clicked. the player then needs to choose
which reinforcement point they're coming from. 
Robert, I've been getting a good idea in my head of what the ui should look like, we should probably 
discuss it and see if we can't get it implemented soon.